+++ kmath-core/src/commonMain/kotlin/space/kscience/kmath/nd/StructureND.kt
@@ -244,6 +244,7 @@ public interface MutableStructureND<T> : StructureND<T> {
  * Set value at specified indices
  */
 @PerformancePitfall
+@Deprecated("")
 public operator fun <T> MutableStructureND<T>.set(vararg index: Int, value: T) {
     set(index, value)
 }